extractor -> definition rename
