+++ engine/flink/components/base-tests/src/test/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -1,1 +1,1 @@
-pl.touk.nussknacker.engine.flink.table.extractor.MockableCatalogFactory
+pl.touk.nussknacker.engine.flink.table.definition.MockableCatalogFactory
